Generate stylesheet from settings
As opposed to having a static stylesheet in the extension. This
will allow users to change/tweak the font settings later on.

Note that this doesn't yet properly apply multiple fonts: it
"carpet bombs" the document with the last font set in the CSS.
Next up should be a list of specific elements to overwrite.