Save axes of new font
Prevously the form would be saved before the axes were processed.
This would be corrected once a different event saved the form
(e.g. pulling slider, changing text inputs), but if you didn't
do this, no axes would be saved.

Timing is now corrected so the form is saved *after* adding the
sliders belonging to the font